cli supports block list

+++ cli/brook/build.sh
@@ -118,6 +118,10 @@ mv brook.ipk ../cli/brook/_/brook_linux_ppc64le.ipk
 
 cd ../cli/brook/
 
-nami release github.com/txthinking/brook $1 _
+if [ "$1" = "dev" ]
+then
+    exit
+fi
 
+nami release github.com/txthinking/brook $1 _
 rm -rf _
